Dynamic content added to frontend

+++ admin/delete-category.php
@@ -5,11 +5,14 @@ if (isset($_GET['id'])) {
     $id = filter_var($_GET['id'], FILTER_SANITIZE_NUMBER_INT);
 
     // Update category of deleted post to uncategorized
+    $update_query = "UPDATE posts SET category_id=5 WHERE category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
 
-
-    // Delete category
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Deleted category successfully";
+    if (!mysqli_errno($connection)) {
+        // Delete category
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Deleted category successfully";
+    }
 }
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
